refactor(app.py): Phase 3 — extract PDF import logic into services/pdf_import.py
Moves all resume/PDF import machinery out of app.py into a new
services/pdf_import.py module: prompt builders, response parsers,
education normalisation, merge helpers, async job state (_pdf_jobs
dict + _pdf_jobs_lock), the ThreadPoolExecutor singleton, TTL/timeout
constants, and the background worker _run_pdf_import_job.

Also updates tests/test_pdf_async.py to rebind _last_prune_time on
the service module rather than on app — app_module rebinding does
not propagate once the attribute is owned by services.pdf_import.
Patch targets for build_provider_chain and generate_with_fallback in
TestRunPdfImportJob updated to target services.pdf_import (the module
that owns the names). PdfReader patch targets in test_profile_import.py
updated to services.pdf_import.PdfReader for the same reason.

Part of #261. Closes #322.
